requirements.txt
numpy==1.14.5
pyserial==3.4
pyserial-asyncio==0.4
PyYAML==3.12
scipy==1.1.0
tornado==5.0.2
websockets==5.0.1
